LayoutTests/webgpu/render-command-encoding-expected.txt

PASS [object WebGPU] is defined.

PASS WebGPURenderPassEncoder created and successfully ended 
